requirements.txt
Flask==2.1.0
mysql-connector-python==8.0.28
python-dotenv==0.19.0
